Update to 1.21.6-rc1

+++ src/main/java/de/maxhenkel/easyvillagers/gui/InputOutputScreen.java
@@ -26,7 +26,7 @@ public abstract class InputOutputScreen<T extends AbstractContainerMenu> extends
     protected void renderLabels(GuiGraphics guiGraphics, int x, int y) {
         drawCentered(guiGraphics, getTopText(), 9, FONT_COLOR);
         drawCentered(guiGraphics, getBottomText(), 40, FONT_COLOR);
-        guiGraphics.drawString(font, playerInventory.getDisplayName().getVisualOrderText(), 8F, (float) (imageHeight - 96 + 3), FONT_COLOR, false);
+        guiGraphics.drawString(font, playerInventory.getDisplayName().getVisualOrderText(), 8, imageHeight - 96 + 3, FONT_COLOR, false);
     }
 
     protected abstract MutableComponent getTopText();
